+++ libs/feature/audit-builder/src/lib/audit-builder/audit-builder.service.ts
@@ -12,7 +12,7 @@ import {
   StepDetails,
   StepProperty, StepType,
 } from '../schema/audit-builder.types';
-import { EMPTY_STEP, STEP_OPTIONS } from '../schema/audit-step.scheme';
+import { EMPTY_STEP, STEP_OPTIONS } from '../schema/audit-step.schema';
 import { STEP_TYPE } from '../schema/step-type.constants';
 import { INPUT_TYPE, PROPERTY_NAME } from '../schema/step-property.constants';
